+++ changelog.md
@@ -7,13 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- [Added support for metadata retrieval](https://github.com/ballerina-platform/ballerina-standard-library/issues/3061)
 
 ### Changed
+- [Fix sql:queryConcat not working for empty query](https://github.com/ballerina-platform/ballerina-standard-library/issues/3127)
 
 ## [1.4.1] - 2022-06-21
 
 ### Changed
-
 - [Fix NullPointerException when retrieving record with default value](https://github.com/ballerina-platform/ballerina-standard-library/issues/2985)
 
 ## [1.4.0] - 2022-05-30