+++ changelog.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Changed
+
+## [1.5.0] - 2022-09-08
+
+### Added
 - [Added support for metadata retrieval](https://github.com/ballerina-platform/ballerina-standard-library/issues/3061)
 
 ### Changed